requirements.txt
tensorflow==2.15.0
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
opencv-python==4.8.0.76
pillow==10.0.0
werkzeug==3.0.0
scikit-learn==1.3.0
matplotlib==3.8.0
wrapt==1.14.1
h5py==3.10.0
gunicorn==21.2.0
